// fema-web-client/src/libs/api-client/model/fMEAType.ts
/**
 * Fema API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const FMEAType = {
    Dfmea: 'DFMEA',
    Pfmea: 'PFMEA'
} as const;
export type FMEAType = typeof FMEAType[keyof typeof FMEAType];